Version_2/requirements.txt
ultralytics
pillow
fastapi
python-multipart
uvicorn
opencv-python-headless
scikit-image